Rename handle_changes to run_queries
